rename workers -> server module
