+++ verifiers/serve/client/zmq_env_client.py
@@ -11,9 +11,9 @@ import zmq.asyncio
 
 from verifiers.utils.logging_utils import print_time
 from verifiers.utils.worker_utils import msgpack_encoder
-from verifiers.workers.client.env_client import EnvClient
-from verifiers.workers.server.zmq_env_server import derive_health_address
-from verifiers.workers.types import (
+from verifiers.serve.client.env_client import EnvClient
+from verifiers.serve.server.zmq_env_server import derive_health_address
+from verifiers.serve.types import (
     BaseRequest,
     BaseResponseT,
     HealthRequest,
