+++ verifiers/serve/server/env_worker.py
@@ -24,7 +24,7 @@ from verifiers.types import ClientConfig
 from verifiers.utils.async_utils import EventLoopLagMonitor
 from verifiers.utils.client_utils import resolve_client_config
 from verifiers.utils.worker_utils import msgpack_encoder
-from verifiers.workers.types import (
+from verifiers.serve.types import (
     BaseResponse,
     EventLoopLagStats,
     RunGroupRequest,
