--- verifiers/workers/server/__init__.py
@@ -1,11 +1,0 @@
-from verifiers.workers.server.env_router import EnvRouter
-from verifiers.workers.server.env_server import EnvServer
-from verifiers.workers.server.env_worker import EnvWorker
-from verifiers.workers.server.zmq_env_server import ZMQEnvServer
-
-__all__ = [
-    "EnvRouter",
-    "EnvServer",
-    "EnvWorker",
-    "ZMQEnvServer",
-]
